mv script to /scripts; add serialasation test for SLT; added optinal toshi_client dependency; refactoting;
