+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 # Changelog
 
-## v1.0.0 - 2022-10-04
+## [0.1.0] - 2022-12-12
+
+## Changed
+ - refactored project structure for packaging
+ - scripts moved to /scripts package
+ - changelog format for bump2version
+
+## Added
+ - poetry with: pytest, coverage, tox, flake8, mypy, black, isort, bump2version
+ - github actions for CI/CD
+ - source_logic_tree feature
+
+## [0.0.0] - 2022-10-04
 Initial release of the NZ NSHM 2022 revision
