+++ pyproject.toml
@@ -2,7 +2,7 @@
 name = "nzshm-model"
 version = "0.1.0"
 description = "The logic tree definitions, final configurations, and versioning of the New Zealand | Aotearoa National Seismic Hazard Model"
-authors = ["Your Name <you@example.com>"]
+authors = ["Chris DiCaprio <c.dicaprio@gns.cri.nz>", "Chris Chamberlain <chrisbc@artisan.co.nz>"]
 license = "AGPL3"
 readme = "README.md"
 packages = [
@@ -12,6 +12,7 @@ packages = [
 
 [tool.poetry.dependencies]
 python = ">=3.8,<4.0"
+nshm-toshi-client = {version = ">=1.0.0", extras = ["toshi"], optional = true}
 
 [tool.poetry.extras]
 test = [
