+++ scripts/convert_LT_2_json.py
