+++ tests/test_logic_tree.py
@@ -1,13 +1,15 @@
 #! python test_logic_tree.py
 
+import dataclasses
+
+import nzshm_model
 from nzshm_model.source_logic_tree.logic_tree import (
+    Branch,
     BranchAttribute,
     BranchAttributeValue,
-    Branch,
     FaultSystemLogicTree,
     SourceLogicTree,
 )
-import itertools
 
 
 def test_direct_bav():
@@ -39,29 +41,10 @@ def test_init():
     assert f.branches[0].values[0].value == 4
 
 
-def build_crustal_branches():
-    # define the branch attributes for our FaultSystemLogicTree
-    C = BranchAttributeValue.all_from_branch_attribute(
-        BranchAttribute(name='C', long_name='area-magnitude scaling', value_options=[4])
-    )
-    s = BranchAttributeValue.all_from_branch_attribute(
-        BranchAttribute(name='s', long_name='moment rate scaling', value_options=[0.5, 1.0])
-    )
-    bN = BranchAttributeValue.all_from_branch_attribute(
-        BranchAttribute(name='bN', long_name='bN pair', value_options=[(0.87, 25), (0.95, 19.2)])
-    )
-    dm = BranchAttributeValue.all_from_branch_attribute(
-        BranchAttribute(name='dm', long_name='deformation model', value_options=['Geodetic', 'Geologic'])
-    )
-
-    crustal_branches = FaultSystemLogicTree('Cru', 'Crustal')
-    for (a, b, c, d) in itertools.product(C, s, bN, dm):
-        crustal_branches.branches.append(Branch(values=[a, b, c, d], weight=0.125, inversion_source='ABC'))
-    return crustal_branches
-
-
 def test_fslt_example():
-    fslt = build_crustal_branches()
+    model_v1_0_0 = nzshm_model.get_model_version('NSHM_1.0.0')['model']
+    fslt = model_v1_0_0.build_crustal_branches()
+
     print(fslt)
     assert fslt.branches[-1].values[0].name == 'C'
     assert fslt.branches[-1].values[0].long_name == 'area-magnitude scaling'
@@ -75,6 +58,19 @@ def test_fslt_example():
 
 
 def test_full_slt():
-    slt = SourceLogicTree(fault_system_branches=[build_crustal_branches()])
+    model_v1_0_0 = nzshm_model.get_model_version('NSHM_1.0.0')['model']
+    slt = SourceLogicTree(fault_system_branches=[model_v1_0_0.build_crustal_branches()])
     print(slt)
     assert slt.fault_system_branches[0].branches[-1].values[0].name == 'C'
+
+
+def test_serialise_slt():
+    model_v1_0_0 = nzshm_model.get_model_version('NSHM_1.0.0')['model']
+    slt = SourceLogicTree(fault_system_branches=[model_v1_0_0.build_crustal_branches()])
+
+    slt_dict = dataclasses.asdict(slt)
+    print(slt_dict)
+    assert (
+        slt.fault_system_branches[0].branches[-1].values[0].name
+        == slt_dict['fault_system_branches'][0]['branches'][-1]['values'][0]['name']
+    )
