+++ tests/test_module_import.py
@@ -17,12 +17,12 @@ class TestStructure:
         assert len(self.model.versions) == 1
 
     def test_version_v1_0_0(self):
-        v1_0_0 = self.model.get_model_version(self.model, 'NSHM_1.0.0')
+        v1_0_0 = self.model.get_model_version('NSHM_1.0.0')
 
         assert v1_0_0['id'] == 'NSHM_1.0.0'
         assert v1_0_0['model'].slt_config is not None
         assert v1_0_0['model'].slt_config.logic_tree_permutations is not None
 
     def test_get_model_version_unknown(self):
-        unknown = self.model.get_model_version(self.model, 'XXX')
+        unknown = self.model.get_model_version('XXX')
         assert unknown is None
